pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "sigis-backend"
version = "0.1.0"
description = "SIGIS — traçabilité missions d'inspection (backend FastAPI, DDD)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    # Auth
    "passlib>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0",
    # Email
    "aiosmtplib>=3.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.29.0",
    "geoalchemy2>=0.14.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
# E501 ignoré : les chaînes de documentation Markdown ne peuvent pas être wrappées à 100 chars.
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["api", "application", "common", "domain", "infrastructure"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
branch = true
source = ["api", "application", "common", "domain", "infrastructure"]
omit = ["*/tests/*", "*/__pycache__/*", "*/alembic/*"]
